Remove setup page & update the index
