+++ installer/README.md
@@ -1,32 +1,22 @@
 # Sedy Installer
-A frontend application to make you install [Sedy](../) easily
 
-# Install npm dependencies
+## Install npm dependencies
 
 ```bash
 make install
 ```
 
-# Run project
+## Run project
 
 ```bash
 make run
 ```
 
-# Configuration
+## Configuration
 
 ```js
 export default {
-    appBaseUrl: 'http://localhost:8080',
-    githubRedirection: '/setup',
-    githubAppId: 'xxx',
-    githubScopes: [
-        'write:repo_hook',
-        'repo',
-    ],
-    githubUrl: 'https://api.github.com',
-    webhookUrl: 'https://sedy.marmelab.com',
-    sedyUsername: 'sedy-bot',
+    ga: 'XXX', // Google Analytics
 };
 
 ```
