+++ installer/config/development-dist.js
@@ -1,6 +1,1 @@
-export default {
-    github: {
-        appId: 'xxxx',
-        redirect_uri: 'http://localhost:8080',
-    },
-};
+export default {};
