+++ installer/config/production-dist.js
@@ -1,15 +1,3 @@
 export default {
-    appBaseUrl: 'https://marmelab.com/sedy',
     ga: 'XXX',
-    setup: 'https://marmelab.com/sedy/setup',
-    github: {
-        url: 'https://github.com',
-        api: 'https://api.github.com',
-        appId: 'xxx',
-        redirect_uri: 'https://marmelab.com/sedy',
-    },
-    oauth: {
-        url: 'http://aws-oauth-lambda',
-        secret: 'MY-VERY-PRIVATE-OAUTH-SECRET',
-    },
 };
