--- installer/js/setup/repositories/getCredentials.js
@@ -1,6 +1,0 @@
-
-export default () => {
-    const { accessToken, user } = window.localStorage;
-
-    return { accessToken, user: user ? JSON.parse(user) : null };
-};
