--- installer/setup.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link href="https://fonts.googleapis.com/css?family=Lato" rel="stylesheet">
-    <title><%= htmlWebpackPlugin.options.title %></title>
-    <meta name="Description" content="<%= htmlWebpackPlugin.options.description %>" />
-    <meta name="Keywords" content="github, pull-requests, typos" />
-
-    <!-- Google Analytics -->
-    <%= htmlWebpackPlugin.options.ga %>
-</head>
-<body>
-    <div id="setup"></div>
-</body>
-</html>
