+++ package.json
@@ -25,5 +25,6 @@
     "eslint-plugin-mocha": "~4.8.0",
     "eslint-plugin-react": "~6.9.0",
     "pm2": "~2.3.0"
-  }
+  },
+  "dependencies": {}
 }
